All features of the plan have been implemented in beta versions.

+++ frontend/src/app/modals/delete-guarantee/delete-guarantee.component.ts
@@ -51,6 +51,10 @@ export class DeleteGuaranteeComponent implements OnInit{
   }
 
   deleteGuarantee(){
+    const confirmDelete = window.confirm("Are you sure?");
+    if (!confirmDelete){
+      return;
+    }
     this.guaranteeService.deleteGuarantee({
       guaranteeId:this.selectedGuaranteeIdToDelete
     }).subscribe({
